separate server and client code
