+++ docs/03-artifacts.md
@@ -147,7 +147,7 @@ The server file processes the document for the artifact. It streams updates (if 
 
 ```ts
 import { smoothStream, streamText } from "ai";
-import { myProvider } from "@/lib/ai/models";
+import { myProvider } from "@/lib/ai/providers";
 import { createDocumentHandler } from "@/lib/artifacts/server";
 import { updateDocumentPrompt } from "@/lib/ai/prompts";
