+++ lib/ai/tools/request-suggestions.ts
@@ -4,7 +4,7 @@ import { DataStreamWriter, streamObject, tool } from 'ai';
 import { getDocumentById, saveSuggestions } from '@/lib/db/queries';
 import { Suggestion } from '@/lib/db/schema';
 import { generateUUID } from '@/lib/utils';
-import { myProvider } from '../models';
+import { myProvider } from '../providers';
 
 interface RequestSuggestionsProps {
   session: Session;
