+++ lib/constants.ts
@@ -1,6 +1,6 @@
-export const isProduction = process.env.NODE_ENV === 'production';
+export const isProductionEnvironment = process.env.NODE_ENV === 'production';
 
-export const isTest = Boolean(
+export const isTestEnvironment = Boolean(
   process.env.PLAYWRIGHT_TEST_BASE_URL ||
     process.env.PLAYWRIGHT ||
     process.env.CI_PLAYWRIGHT,
